ref: move WebhooksSection to own directory
