+++ frontend/src/features/admin-form/settings/components/WebhooksSection/index.ts
@@ -1,0 +1,1 @@
+export { WebhooksSection } from './WebhooksSection'
